Benchmark differences
```
goos: linux
goarch: amd64
pkg: github.com/push-and-pray/btree
cpu: 12th Gen Intel(R) Core(TM) i5-12400F
                                            │   old.txt    │               new.txt               │
                                            │    sec/op    │    sec/op     vs base               │
BTreeInsert/Insert_Degree_2_Size_+1000-12     97.78µ ±  1%   90.36µ ±  2%  -7.58% (p=0.000 n=10)
BTreeInsert/Insert_Degree_2_Size_+10000-12    1.799m ±  1%   1.744m ±  1%  -3.05% (p=0.000 n=10)
BTreeInsert/Insert_Degree_2_Size_+100000-12   38.03m ±  3%   35.59m ±  2%  -6.41% (p=0.000 n=10)
BTreeInsert/Insert_Degree_4_Size_+1000-12     78.50µ ± 10%   72.60µ ±  1%       ~ (p=0.436 n=10)
BTreeInsert/Insert_Degree_4_Size_+10000-12    1.313m ±  1%   1.269m ±  2%  -3.28% (p=0.000 n=10)
BTreeInsert/Insert_Degree_4_Size_+100000-12   22.23m ±  4%   21.79m ±  4%       ~ (p=0.063 n=10)
BTreeInsert/Insert_Degree_8_Size_+1000-12     66.44µ ±  1%   59.92µ ±  1%  -9.81% (p=0.000 n=10)
BTreeInsert/Insert_Degree_8_Size_+10000-12    1.159m ±  1%   1.118m ±  1%  -3.58% (p=0.000 n=10)
BTreeInsert/Insert_Degree_8_Size_+100000-12   18.19m ±  1%   17.86m ±  1%  -1.80% (p=0.002 n=10)
BTreeGet/Get_Degree_2_Size_+1000-12           89.24µ ±  2%   83.53µ ±  2%  -6.40% (p=0.000 n=10)
BTreeGet/Get_Degree_2_Size_+10000-12          1.718m ±  2%   1.675m ±  2%  -2.55% (p=0.001 n=10)
BTreeGet/Get_Degree_2_Size_+100000-12         35.42m ±  3%   34.19m ±  2%  -3.47% (p=0.000 n=10)
BTreeGet/Get_Degree_4_Size_+1000-12           74.08µ ±  9%   66.94µ ± 11%  -9.65% (p=0.002 n=10)
BTreeGet/Get_Degree_4_Size_+10000-12          1.257m ±  1%   1.233m ±  2%  -1.92% (p=0.025 n=10)
BTreeGet/Get_Degree_4_Size_+100000-12         21.15m ±  3%   20.75m ±  2%  -1.93% (p=0.029 n=10)
BTreeGet/Get_Degree_8_Size_+1000-12           61.93µ ±  1%   56.42µ ±  2%  -8.89% (p=0.000 n=10)
BTreeGet/Get_Degree_8_Size_+10000-12          1.110m ±  2%   1.087m ±  1%  -2.11% (p=0.002 n=10)
BTreeGet/Get_Degree_8_Size_+100000-12         17.49m ±  1%   17.40m ±  1%       ~ (p=0.315 n=10)
BTreeDelete/Delete_Degree_2_Size_+1000-12     1.059µ ±  1%   1.036µ ±  1%  -2.17% (p=0.000 n=10)
BTreeDelete/Delete_Degree_2_Size_+10000-12    10.44µ ±  1%   10.32µ ±  1%  -1.13% (p=0.000 n=10)
BTreeDelete/Delete_Degree_2_Size_+100000-12   105.8µ ±  1%   106.0µ ±  1%       ~ (p=0.739 n=10)
BTreeDelete/Delete_Degree_4_Size_+1000-12     1.047µ ±  1%   1.037µ ±  1%  -0.96% (p=0.015 n=10)
BTreeDelete/Delete_Degree_4_Size_+10000-12    10.39µ ±  1%   10.36µ ±  1%       ~ (p=0.247 n=10)
BTreeDelete/Delete_Degree_4_Size_+100000-12   106.0µ ±  1%   105.3µ ±  1%  -0.62% (p=0.001 n=10)
BTreeDelete/Delete_Degree_8_Size_+1000-12     1.052µ ±  1%   1.035µ ±  1%  -1.62% (p=0.000 n=10)
BTreeDelete/Delete_Degree_8_Size_+10000-12    10.42µ ±  1%   10.40µ ±  1%       ~ (p=0.247 n=10)
BTreeDelete/Delete_Degree_8_Size_+100000-12   106.3µ ±  1%   104.3µ ±  1%  -1.86% (p=0.000 n=10)
geomean                                       269.8µ         260.6µ        -3.42%

                                            │     old.txt      │                 new.txt                 │
                                            │       B/op       │     B/op       vs base                  │
BTreeInsert/Insert_Degree_2_Size_+1000-12        7.000 ± 14%       7.000 ±  0%        ~ (p=0.474 n=10)
BTreeInsert/Insert_Degree_2_Size_+10000-12     1.416Ki ±  1%     1.381Ki ±  2%   -2.45% (p=0.002 n=10)
BTreeInsert/Insert_Degree_2_Size_+100000-12    327.8Ki ±  9%     282.9Ki ±  3%  -13.69% (p=0.000 n=10)
BTreeInsert/Insert_Degree_4_Size_+1000-12        3.000 ±  0%       3.000 ± 33%        ~ (p=0.474 n=10)
BTreeInsert/Insert_Degree_4_Size_+10000-12       610.0 ±  1%       586.5 ±  2%   -3.85% (p=0.000 n=10)
BTreeInsert/Insert_Degree_4_Size_+100000-12   116.61Ki ± 15%     97.64Ki ± 10%  -16.26% (p=0.004 n=10)
BTreeInsert/Insert_Degree_8_Size_+1000-12        2.000 ±  0%       2.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeInsert/Insert_Degree_8_Size_+10000-12       423.0 ±  3%       406.5 ±  3%   -3.90% (p=0.002 n=10)
BTreeInsert/Insert_Degree_8_Size_+100000-12    71.69Ki ±  5%     71.65Ki ± 12%        ~ (p=0.739 n=10)
BTreeGet/Get_Degree_2_Size_+1000-12              0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_2_Size_+10000-12             0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_2_Size_+100000-12            0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_4_Size_+1000-12              0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_4_Size_+10000-12             0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_4_Size_+100000-12            0.000 ±   ?       0.000 ±   ?        ~ (p=1.000 n=10)
BTreeGet/Get_Degree_8_Size_+1000-12              0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_8_Size_+10000-12             0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_8_Size_+100000-12            0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_2_Size_+1000-12        0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_2_Size_+10000-12       0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_2_Size_+100000-12      0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_4_Size_+1000-12        0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_4_Size_+10000-12       0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_4_Size_+100000-12      0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_8_Size_+1000-12        0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_8_Size_+10000-12       0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_8_Size_+100000-12      0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
geomean                                                      ²                   -1.58%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                            │    old.txt     │                new.txt                 │
                                            │   allocs/op    │  allocs/op    vs base                  │
BTreeInsert/Insert_Degree_2_Size_+1000-12      0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeInsert/Insert_Degree_2_Size_+10000-12     33.00 ±  3%      32.50 ±  2%   -1.52% (p=0.005 n=10)
BTreeInsert/Insert_Degree_2_Size_+100000-12   7.866k ±  9%     6.790k ±  3%  -13.69% (p=0.000 n=10)
BTreeInsert/Insert_Degree_4_Size_+1000-12      0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeInsert/Insert_Degree_4_Size_+10000-12     8.000 ±  0%      7.000 ± 14%  -12.50% (p=0.020 n=10)
BTreeInsert/Insert_Degree_4_Size_+100000-12   1.599k ± 15%     1.339k ± 10%  -16.27% (p=0.004 n=10)
BTreeInsert/Insert_Degree_8_Size_+1000-12      0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeInsert/Insert_Degree_8_Size_+10000-12     3.000 ±  0%      2.000 ± 50%  -33.33% (p=0.005 n=10)
BTreeInsert/Insert_Degree_8_Size_+100000-12    528.5 ±  5%      528.5 ± 12%        ~ (p=0.753 n=10)
BTreeGet/Get_Degree_2_Size_+1000-12            0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_2_Size_+10000-12           0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_2_Size_+100000-12          0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_4_Size_+1000-12            0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_4_Size_+10000-12           0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_4_Size_+100000-12          0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_8_Size_+1000-12            0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_8_Size_+10000-12           0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeGet/Get_Degree_8_Size_+100000-12          0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_2_Size_+1000-12      0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_2_Size_+10000-12     0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_2_Size_+100000-12    0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_4_Size_+1000-12      0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_4_Size_+10000-12     0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_4_Size_+100000-12    0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_8_Size_+1000-12      0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_8_Size_+10000-12     0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
BTreeDelete/Delete_Degree_8_Size_+100000-12    0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=10) ¹
geomean                                                    ²                  -3.20%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```